requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
astroid==2.4.2
asttokens==2.0.5
attrs==21.2.0
backcall==0.2.0
black==19.10b0
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
colorlog==4.8.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
elasticsearch==7.11.0
entrypoints==0.4
enum34==1.1.10
executing==0.8.3
fake-useragent==0.1.11
filetype==1.0.7
Flask==2.1.1
grpcio==1.38.0
gunicorn==20.1.0
h11==0.12.0
h2==2.6.2
har2case==0.3.1
hpack==3.0.0
httprunner==2.5.7
httptools==0.1.1
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.8.1
iniconfig==1.1.1
ipykernel==5.4.0
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.17.2
Jinja2==3.1.1
jira==3.0.1
jmespath==0.9.5
jsonpath==0.82
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
keyring==23.2.1
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
loguru==0.4.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
more-itertools==8.8.0
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.19.3
oauthlib==3.1.1
opencv-contrib-python==4.5.4.58
packaging==21.0
parso==0.7.1
pathspec==0.9.0
pbr==5.6.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pluggy==0.13.1
ply==3.11
prompt-toolkit==3.0.8
protobuf==3.17.1
ptyprocess==0.6.0
pure-eval==0.2.2
py==1.10.0
pycryptodome==3.10.1
pydantic==1.8.2
Pygments==2.7.3
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==7.1.2
pytest-html==2.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
redis==3.5.3
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
scikit-image==0.18.3
scipy==1.7.1
sentry-sdk==0.13.5
six==1.15.0
sqlparse==0.4.1
stack-data==0.2.0
tifffile==2021.8.30
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.1
wcwidth==0.2.5
wd==1.0.1
Werkzeug==2.1.1
What==0.5.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
XmindExcelTestcases==0.0.8
xmindparser==1.0.9
zipp==3.6.0
